+++ mystays/models.py
@@ -15,7 +15,7 @@ class Stay(models.Model):
     name = models.CharField(max_length=128, unique=True)
 #    stayID = models.CharField(max_length=30, unique=True)
     picture = models.ImageField(upload_to='stay_images/', default='default.jpg')
-    price = models.IntegerField(default=0)
+    price = models.DecimalField(max_digits=10, decimal_places=2, default=0)
     propertyRating = models.DecimalField(max_digits=3, decimal_places=1, default=0.0)
     description = models.CharField(max_length=255)
     latitude = models.DecimalField(max_digits=15, decimal_places=10, default=0)
